setup testing, CI
- Setup a worflow for running pytest unit tests
- Move main app into app.py
- Add CI that runs pytest
- Add pytest dependency
