+++ .github/workflows/pytest.yml
@@ -1,0 +1,26 @@
+name: pytest
+
+on:
+  push:
+    branches:
+      - "*"
+
+jobs:
+  backend-unit-tests:
+    name: python
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v5
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v7
+        with:
+          enable-cache: true
+
+      - name: Install the project
+        run: uv sync --locked --all-extras --dev
+
+      - name: Run tests
+
+        run: uv run pytest
