+++ pyproject.toml
@@ -9,6 +9,8 @@ authors = [
 requires-python = ">=3.13"
 dependencies = [
     "fastapi>=0.123.0",
+    "httpx>=0.28.1",
+    "pytest>=9.0.1",
     "sqlalchemy>=2.0.44",
     "uvicorn>=0.38.0",
 ]
